+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 # 4.46.1
 - Fix Android app crash on old Android versions
+- Fix Linux blank screen issue related to the local mimetype database
 
 # 4.46.0
 - Android: enable export logs feature